[CAN 드라이버 오류 수정] - generateJ1939FrameBuffer() 함수에서 source address가 uint16으로 되어 있는 부분을 uint8로 수정

+++ yss/drv/can/drv_can_stm32f1_f4_f7.cpp
@@ -310,7 +310,7 @@ uint8_t Can::getReceiveErrorCount(void)
 	return (mPeri[ERR] >> 24);
 }
 
-J1939Frame Can::generateJ1939FrameBuffer(uint8_t priority, uint16_t pgn, uint16_t sa, uint8_t count)
+J1939Frame Can::generateJ1939FrameBuffer(uint8_t priority, uint16_t pgn, uint8_t sa, uint8_t count)
 {
 	J1939Frame buf = {0, 0, true, sa, pgn, 0, 0, priority, count, 0, 0,};
 	return buf;
